+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v3.1.0-dev (Aug 08, 2024)
-
 - Add `defaultDnt` to support setting the dnt flag for SLAS. Upgrade `commerce-sdk-isomorphic` to v3.1.1 [#1979](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1979)
+- Added useDNT hook to commerce-sdk-react [#1989](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1989)
 - Read the DNT preference from the `dw_dnt` cookie set by the `useDnt` hook. If the cookie is not set, `defaultDnt` will be used to set the DNT flag. [#1987] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1987)
 - Update logout helper to work for guest users [#1997](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1997)